refactor/bug: moved search logic; fix quoted words
